Refactor JWT
