+++ src/components/Admin/AdminDashboard/SenseiGrowthChart/index.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react'
 import ChartistGraph from 'react-chartist'
-import * as jwtAdmin from 'services/jwt/admin'
+import * as jwtAdmin from 'services/admin'
 import ChartistTooltip from 'chartist-plugin-tooltips-updated'
 import { MONTH_NAMES } from 'constants/constants'
