+++ src/components/Admin/AdminManagement/AdminProfile/index.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react'
 import { useParams, useHistory } from 'react-router-dom'
 import { useSelector } from 'react-redux'
 import { Button, Descriptions, Modal, Form, Input, Select, Popconfirm } from 'antd'
-import * as jwtAdmin from 'services/jwt/admin'
+import * as jwtAdmin from 'services/admin'
 import { EditOutlined, DeleteOutlined, ArrowLeftOutlined } from '@ant-design/icons'
 import { ADMIN_ROLE_ENUM } from 'constants/constants'
 import { showNotification } from 'components/utils'
