+++ src/components/Admin/AdminManagement/AdminTable/index.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react'
 import { useHistory } from 'react-router-dom'
 import { useSelector } from 'react-redux'
 import { Button, Tabs, Table } from 'antd'
-import * as jwtAdmin from 'services/jwt/admin'
+import * as jwtAdmin from 'services/admin'
 import { InfoCircleOutlined } from '@ant-design/icons'
 import { indexOf, isNil, map } from 'lodash'
 import { formatTime } from 'components/utils'
