+++ src/components/Admin/AdminManagement/SuperAdminWidget/index.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react'
 import CountIconWidget from 'components/Common/CountIconWidget'
 import { SafetyCertificateFilled } from '@ant-design/icons'
 import { isNil, size } from 'lodash'
-import * as jwtAdmin from 'services/jwt/admin'
+import * as jwtAdmin from 'services/admin'
 import { ADMIN_ROLE_ENUM } from 'constants/constants'
 
 const SuperAdminWidget = () => {
