+++ src/components/Admin/Mentorship/ContractsWidget/index.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react'
-import * as jwtAdmin from 'services/jwt/admin'
+import * as jwtAdmin from 'services/admin'
 
 const ContractsWidget = () => {
   const [count, setCount] = useState(0)
