+++ src/redux/user/sagas.js
@@ -3,7 +3,7 @@ import { notification } from 'antd'
 import { history } from 'index'
 import { USER_TYPE_ENUM } from 'constants/constants'
 import { isEmpty, isNil } from 'lodash'
-import * as jwt from 'services/jwt'
+import * as jwt from 'services/user'
 import { createAdminObj, createUserObj, resetCart, resetUser } from 'components/utils'
 import actions from './actions'
 import * as selectors from '../selectors'
